Release 1.4.13

+++ worldlineop.php
@@ -17,6 +17,7 @@ if (!defined('_PS_VERSION_')) {
 
 use Monolog\Logger;
 use PrestaShop\ModuleLibServiceContainer\DependencyInjection\ServiceContainer;
+use WorldlineOP\PrestaShop\Utils\Tools as ToolsWorldline;
 
 /**
  * Class Worldlineop
@@ -41,7 +42,7 @@ class Worldlineop extends PaymentModule
 
         $this->name = 'worldlineop';
         $this->author = 'Worldline Online Payments';
-        $this->version = '1.4.12';
+        $this->version = '1.4.13';
         $this->tab = 'payments_gateways';
         $this->module_key = '089d13d0218de8085259e542483f4438';
         $this->currencies = true;
@@ -95,11 +96,24 @@ class Worldlineop extends PaymentModule
      */
     public function uninstall()
     {
-        Configuration::deleteByName('WORLDLINEOP_ACCOUNT_SETTINGS');
-        Configuration::deleteByName('WORLDLINEOP_ADVANCED_SETTINGS');
-        Configuration::deleteByName('WORLDLINEOP_PAYMENT_METHODS_SETTINGS');
+        if (parent::uninstall()) {
+            Configuration::deleteByName('WORLDLINEOP_ACCOUNT_SETTINGS');
+            Configuration::deleteByName('WORLDLINEOP_ADVANCED_SETTINGS');
+            Configuration::deleteByName('WORLDLINEOP_PAYMENT_METHODS_SETTINGS');
+            ToolsWorldline::removeSymfonyCache();
+            return true;
+        }
 
-        return parent::uninstall();
+        return false;
+    }
+
+    public function disable($force_all = false)
+    {
+        if (parent::disable($force_all)) {
+            ToolsWorldline::removeSymfonyCache();
+            return true;
+        }
+        return false;
     }
 
     /**
